add module code without destroy_store
