+++ driver/src/module/main.c
@@ -1,4 +1,6 @@
 #include "../cremona/central_store.h"
+#include "module.h"
+#include "net_link_comm.h"
 #include <linux/module.h>
 #include <linux/netlink.h>
 char *crmna_prefix;
@@ -9,30 +11,25 @@ static const unsigned int MINOR_BASE = 0;
 static const unsigned int MINOR_NUM = 33; /* マイナー番号は 0 ~ 1 */
 
 store_t *store;
-
-static int netlink_receiver(struct sk_buff *skb, struct nlmsghdr *nlh,
-                            struct netlink_ext_ack *ack) {
-  DEFINE_ERROR(err);
-  DEFINE_CRMNA_BUF_FROM_MEMORY(data, (char *)NLMSG_DATA(nlh), nlh->nlmsg_len);
-  action_t action;
-  create_action_from_message(nlh->nlmsg_type, nlh->nlmsg_pid, &data, &action, &err);
-  return 0;
-}
+communicator_ref_t com;
 
 static int cremona_init(void) {
   printk("Hello my module\n");
 
-  store = create_store();
-
-  // if (!create_device_manager_cntext(&context,MINOR_BASE, MINOR_NUM, DRIVER_NAME, "cremona", 17)) {
-  //   printk("Error in create device manager context.\n");
-  //   return -ENOENT;
-  // }
+  store = create_store(MINOR_BASE, MINOR_NUM, DRIVER_NAME);
+  if(store == NULL){
+    return -EFAULT;
+  }
+  if(!create_communicator_and_register_store(17, store, &ref)){
+    destroy_store(store);
+    return -EFAULT;
+  }
   return 0;
 }
 
 static void cremona_exit(void) {
-  //destroy_device_manager_context(&context);
+  destroy_store(store);
+  communicator_free(&ref);
   printk("Bye bye my module\n");
 }
